Add note content area
Renders two parts: note sections, and an empty space filling the rest of
the area.

The purpose of the empty space is to add margin space between the note
content and the bottom of the page. I find it annoying when the majority
of my typing is only a few pixels from the bottom of the page. It also
serves as a click handler. If you click after the note content, it
should either focus the last text area or create a new one.