Consolidate `jest-react` and `test-utils` packages
Now there's a new package called `test-utils` that has some enzyme
renderer helper functions and the jest preset from `@ponder/jest-react`
all in one. Everything's using the new package now. This is mainly for
`@ponder/ui`.

I think that's the last of the things I need to make a package for.
Phew.