Add `Input` component
Applies a little bit of style and a lotta bit of logic. It'll only fire
change events after you finish editing, unlike `onChange(...)`. This is
something I've been wanting to write for a long time... a pattern that
allows for controlled inputs without damaging performance through Redux.
This allows you to sync with redux after the user is finished editing.