Select text when inputs are focused
Specifically the note title. When you click on the input, select all the
text so it's easy to replace the title entirely.

Worth noting that this required adding another field to the
`renderer(...)` test utility, something which I was planning anyway. It
allows developers to perform some task whenever the component is
constructed, like setting a ref, which is how I was using it.

Also worth noting: there was a Flow<generic> battle and I lost. This
manifested in the introduction of an `Object` type. I just couldn't
figure out how to tell Flow that a return value of an object parameter's
method should be merged with the parent return value's derived state,
but only if the optional method existed.

Now that I actually say that, it doesn't seem like that big a deal.