Add markdown editor component
That was fun! I got CodeMirror integrated with Ponder. Right now I'm
doing a pretty pointless thing and using a controlled input as an
uncontrolled input, but it'll need to be controlled at some point or
another in order to work with real time updates.

Oh and another thing, I added a Jest mock for any file ending with
`.css` so Jest won't die horribly.