Update emotion
Switches to the latest version of emotion. I have to say, I'm not loving
this library. I kinda regret switching from styled-components. The only
downside of staying would've been not knowing how bad the experience is.

Things that bother me: they seem to make breaking changes often. Every
time I update, there are codebase-wide breaking changes. The other thing
is having to switch to `mount(...)` for _any_ of my tests. I consider
that a heavy anti-pattern, and the fact that `.toHaveStyleRule(...)`
implicitly enforces it makes me want to vomit.