Fix lint
Disables a rule that I personally disagree with. The rule reports places
where I'm depending on inherited object methods because
`Object.create(null)` is a thing. Arguably though, 90% of the time if
I need `Object.create(null)`, `Map` is a better suited alternative. For
the other 10% of the time you already know it's a null proto object. You
deserve the bugs.

But all this is a moot point because Flow doesn't support null proto
objects. God I hate Flow.