Include sections in notebook listings
This is a symptom of a larger problem: sections shouldn't be members of
the `Note` type, rather they should have their own files. The attempt to
read the notebook will read every single note within it as well. Not
great.

The type was complaining because notes should have sections but the
notes returned from `notebook.list()` didn't have 'em.

I might switch to IndexedDB in the future and try to fix that problem,
but hey, the app should be usable before I go full Architecture
Astronaut.